+++ react/components/FlexLayoutContext.tsx
@@ -1,4 +1,4 @@
-import React from 'react'
+import React, { FunctionComponent, useContext, useMemo } from 'react'
 
 export enum FlexLayoutTypes {
   NONE = 'none',
@@ -6,10 +6,41 @@ export enum FlexLayoutTypes {
   COL = 'col',
 }
 
-interface FlexLayoutContext {
+export interface FlexLayoutContextProps {
   parent: FlexLayoutTypes
+  colGap: number
+  rowGap: number
 }
 
-export const FlexLayoutContext = React.createContext<FlexLayoutContext>({
+export type FlexLayoutContextValue = number | FlexLayoutTypes
+
+export type FlexLayoutContext = FlexLayoutContextProps & {
+  [key: string]: FlexLayoutContextValue
+}
+
+const FlexLayoutContext = React.createContext<FlexLayoutContext>({
   parent: FlexLayoutTypes.NONE,
+  colGap: 0,
+  rowGap: 0,
 })
+
+export const FlexLayoutContextProvider: FunctionComponent<
+  FlexLayoutContextProps
+> = ({ parent, colGap, rowGap, children }) => {
+  const value = useMemo(
+    () => ({
+      parent,
+      colGap,
+      rowGap,
+    }),
+    [parent, colGap, rowGap]
+  )
+
+  return (
+    <FlexLayoutContext.Provider value={value}>
+      {children}
+    </FlexLayoutContext.Provider>
+  )
+}
+
+export const useFlexLayoutContext = () => useContext(FlexLayoutContext)