+++ CHANGELOG.md
@@ -7,6 +7,17 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.5.1] - 2019-05-23
+### Fixed
+- Fixed missing `id` from schemas.
+
+## [0.5.0] - 2019-05-23
+### Added
+- Props `colGap`, `rowGap`, to add spacing between columns and rows.
+- Prop `preserveLayoutOnMobile`, to refrain from breaking columns into rows on small screens.
+- **Col:** Props `width`, `marginLeft/Right`, `paddingLeft/Right`, and `preventVerticalStretch`.
+- **Row:** Props `marginTop/Bottom`, `paddingTop/Bottom`, `preventHorizontalStretch`, and `preventVerticalStretch`.
+
 ## [0.4.2] - 2019-04-26
 
 ### Fixed